feat: page.js 코드 분리(간소화)
+ 깃허브 이슈 #75
