+++ frontend/src/Component/Page/utils/pageSettings.js
@@ -1,0 +1,29 @@
+import { defaultSettings } from "prosemirror-image-plugin";
+
+export const imageSettings = {
+  ...defaultSettings,
+  hasTitle: false,
+  minSize: 30,
+  maxSize: 550,
+};
+
+export const imageNodeSpec = {
+  inline: false,
+  group: "block",
+  attrs: {
+    src: {},
+    alt: { default: null },
+    title: { default: null },
+  },
+  parseDOM: [
+    {
+      tag: "img[src]",
+      getAttrs: (dom) => ({
+        src: dom.getAttribute("src"),
+        alt: dom.getAttribute("alt"),
+        title: dom.getAttribute("title"),
+      }),
+    },
+  ],
+  toDOM: (node) => ["img", node.attrs],
+};
